move to resilience_config
